[refactoring] step3: 책임 분리
